[DeadCode] Make private property removal dependent on local property fetches
