--- packages/NodeCollector/NodeCollector/ParsedPropertyFetchNodeCollector.php
@@ -1,97 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Rector\NodeCollector\NodeCollector;
-
-use PhpParser\Node;
-use PhpParser\Node\Expr\MethodCall;
-use PhpParser\Node\Expr\PropertyFetch;
-use PhpParser\Node\Expr\StaticCall;
-use PhpParser\Node\Expr\StaticPropertyFetch;
-use PHPStan\Type\MixedType;
-use PHPStan\Type\Type;
-use PHPStan\Type\TypeWithClassName;
-use PHPStan\Type\UnionType;
-use Rector\NodeNameResolver\NodeNameResolver;
-use Rector\NodeTypeResolver\NodeTypeResolver;
-
-/**
- * All parsed nodes grouped type
- */
-final class ParsedPropertyFetchNodeCollector
-{
-    /**
-     * @var array<string, array<string, PropertyFetch[]>>
-     */
-    private array $propertyFetchesByTypeAndName = [];
-
-    public function __construct(
-        private NodeNameResolver $nodeNameResolver,
-        private NodeTypeResolver $nodeTypeResolver
-    ) {
-    }
-
-    public function collect(Node $node): void
-    {
-        if (! $node instanceof PropertyFetch && ! $node instanceof StaticPropertyFetch) {
-            return;
-        }
-
-        $propertyType = $this->resolvePropertyCallerType($node);
-        if ($propertyType instanceof MixedType) {
-            return;
-        }
-
-        // make sure name is valid
-        if ($node->name instanceof StaticCall || $node->name instanceof MethodCall) {
-            return;
-        }
-
-        $propertyName = $this->nodeNameResolver->getName($node->name);
-        if ($propertyName === null) {
-            return;
-        }
-
-        $this->addPropertyFetchWithTypeAndName($propertyType, $node, $propertyName);
-    }
-
-    /**
-     * @return PropertyFetch[]
-     */
-    public function findPropertyFetchesByTypeAndName(string $className, string $propertyName): array
-    {
-        return $this->propertyFetchesByTypeAndName[$className][$propertyName] ?? [];
-    }
-
-    /**
-     * @param PropertyFetch|StaticPropertyFetch $node
-     */
-    private function resolvePropertyCallerType(Node $node): Type
-    {
-        if ($node instanceof PropertyFetch) {
-            return $this->nodeTypeResolver->resolve($node->var);
-        }
-
-        return $this->nodeTypeResolver->resolve($node->class);
-    }
-
-    /**
-     * @param PropertyFetch|StaticPropertyFetch $propertyFetchNode
-     */
-    private function addPropertyFetchWithTypeAndName(
-        Type $propertyType,
-        Node $propertyFetchNode,
-        string $propertyName
-    ): void {
-        if ($propertyType instanceof TypeWithClassName) {
-            $this->propertyFetchesByTypeAndName[$propertyType->getClassName()][$propertyName][] = $propertyFetchNode;
-        }
-
-        if ($propertyType instanceof UnionType) {
-            foreach ($propertyType->getTypes() as $unionedType) {
-                $this->addPropertyFetchWithTypeAndName($unionedType, $propertyFetchNode, $propertyName);
-            }
-        }
-    }
-}
