+++ packages/ReadWrite/NodeAnalyzer/ReadWritePropertyAnalyzer.php
@@ -32,13 +32,8 @@ final class ReadWritePropertyAnalyzer
     ) {
     }
 
-    /**
-     * @param PropertyFetch|StaticPropertyFetch $node
-     */
-    public function isRead(Node $node): bool
+    public function isRead(PropertyFetch|StaticPropertyFetch $node): bool
     {
-        Assert::isAnyOf($node, [PropertyFetch::class, StaticPropertyFetch::class]);
-
         $parent = $node->getAttribute(AttributeKey::PARENT_NODE);
         if (! $parent instanceof Node) {
             throw new ShouldNotHappenException();
