+++ packages/ReadWrite/NodeFinder/NodeUsageFinder.php
@@ -6,12 +6,10 @@ namespace Rector\ReadWrite\NodeFinder;
 
 use PhpParser\Node;
 use PhpParser\Node\Expr;
-use PhpParser\Node\Expr\PropertyFetch;
 use PhpParser\Node\Expr\Variable;
 use PhpParser\Node\Stmt\Foreach_;
 use Rector\Core\PhpParser\Comparing\NodeComparator;
 use Rector\Core\PhpParser\Node\BetterNodeFinder;
-use Rector\NodeCollector\NodeCollector\NodeRepository;
 use Rector\NodeNameResolver\NodeNameResolver;
 use Rector\NodeNestingScope\NodeFinder\ScopeAwareNodeFinder;
 
@@ -20,7 +18,6 @@ final class NodeUsageFinder
     public function __construct(
         private NodeNameResolver $nodeNameResolver,
         private BetterNodeFinder $betterNodeFinder,
-        private NodeRepository $nodeRepository,
         private ScopeAwareNodeFinder $scopeAwareNodeFinder,
         private NodeComparator $nodeComparator
     ) {
@@ -48,25 +45,6 @@ final class NodeUsageFinder
 
             return $this->nodeNameResolver->isName($node, $variableName);
         });
-    }
-
-    /**
-     * @return PropertyFetch[]
-     */
-    public function findPropertyFetchUsages(PropertyFetch $desiredPropertyFetch): array
-    {
-        $propertyFetches = $this->nodeRepository->findPropertyFetchesByPropertyFetch($desiredPropertyFetch);
-
-        $propertyFetchesWithoutPropertyFetch = [];
-        foreach ($propertyFetches as $propertyFetch) {
-            if ($propertyFetch === $desiredPropertyFetch) {
-                continue;
-            }
-
-            $propertyFetchesWithoutPropertyFetch[] = $propertyFetch;
-        }
-
-        return $propertyFetchesWithoutPropertyFetch;
     }
 
     public function findPreviousForeachNodeUsage(Foreach_ $foreach, Expr $expr): ?Node
