+++ rules/DeadCode/Rector/Property/RemoveUnusedPrivatePropertyRector.php
@@ -10,6 +10,7 @@ use PhpParser\Node\Stmt\Interface_;
 use PhpParser\Node\Stmt\Property;
 use PhpParser\Node\Stmt\Trait_;
 use Rector\Core\NodeManipulator\PropertyManipulator;
+use Rector\Core\PhpParser\NodeFinder\PropertyFetchFinder;
 use Rector\Core\Rector\AbstractRector;
 use Rector\NodeTypeResolver\Node\AttributeKey;
 use Rector\Removing\NodeManipulator\ComplexNodeRemover;
@@ -23,7 +24,7 @@ final class RemoveUnusedPrivatePropertyRector extends AbstractRector
 {
     public function __construct(
         private PropertyManipulator $propertyManipulator,
-        private ComplexNodeRemover $complexNodeRemover
+        private ComplexNodeRemover $complexNodeRemover,
     ) {
     }
